+++ apps/explorer/test/test_helper.exs
@@ -16,6 +16,7 @@ Ecto.Adapters.SQL.Sandbox.mode(Explorer.Repo.Account, :auto)
 Ecto.Adapters.SQL.Sandbox.mode(Explorer.Repo.PolygonEdge, :auto)
 Ecto.Adapters.SQL.Sandbox.mode(Explorer.Repo.PolygonZkevm, :auto)
 Ecto.Adapters.SQL.Sandbox.mode(Explorer.Repo.RSK, :auto)
+Ecto.Adapters.SQL.Sandbox.mode(Explorer.Repo.Shibarium, :auto)
 Ecto.Adapters.SQL.Sandbox.mode(Explorer.Repo.Suave, :auto)
 
 Mox.defmock(Explorer.ExchangeRates.Source.TestSource, for: Explorer.ExchangeRates.Source)